requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.13.4
boto3==1.38.18
botocore==1.38.18
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.43.0
dj-database-url==2.3.0
Django==5.1.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==25.1
django-heroku==0.3.1
django-sslserver==0.22
django-storages==1.14.6
djangorestframework==3.15.2
gdacs-api==2.0.0
geographiclib==2.0
geopy==2.4.1
google-auth==2.38.0
google-genai==0.8.0
google_search_results==2.4.2
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
joblib==1.4.2
lxml==5.3.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0a1
pydantic_core==2.28.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
rsa==4.9
s3transfer==0.12.0
scipy==1.15.2
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==14.2
whitenoise==6.8.2
xmltodict==0.14.2
